Refactoring, and remove confidence logging for questions not marked with `showConfidence` attribute
